editcap: Support for a large number of frames
There's nothing in the wiretap packet reading or writing code,
as used by editcap, that cares about the packet number and thus
there's no need to limit editcap to a 32 bit number of frames.

Indeed, if one has such an enormous file, an obvious thing to
try to do is to use editcap to split it into sizes that can be
handled by tshark (2^32 - 1) or Wireshark (2^31 - 1).

Fix #16122

+++ ui/alert_box.h
@@ -72,7 +72,7 @@ extern void cfile_read_failure_alert_box(const char *filename, int err,
 extern void cfile_write_failure_alert_box(const char *in_filename,
                                           const char *out_filename,
                                           int err, char *err_info,
-                                          uint32_t framenum,
+                                          uint64_t framenum,
                                           int file_type_subtype);
 
 /*
