+++ ui/failure_message.h
@@ -92,7 +92,7 @@ extern void cfile_read_failure_message(const char *filename, int err,
 extern void cfile_write_failure_message(const char *in_filename,
                                         const char *out_filename,
                                         int err, char *err_info,
-                                        uint32_t framenum,
+                                        uint64_t framenum,
                                         int file_type_subtype);
 
 /*
