+++ wsutil/report_message.h
@@ -39,7 +39,7 @@ struct report_message_routines {
 	void (*report_cfile_dump_open_failure)(const char *, int, char *, int);
 	void (*report_cfile_read_failure)(const char *, int, char *);
 	void (*report_cfile_write_failure)(const char *, const char *,
-	    int, char *, uint32_t, int);
+	    int, char *, uint64_t, int);
 	void (*report_cfile_close_failure)(const char *, int, char *);
 };
 
@@ -99,7 +99,7 @@ WS_DLL_PUBLIC void report_cfile_read_failure(const char *filename,
  * Report an error from attempting to write to a capture file.
  */
 WS_DLL_PUBLIC void report_cfile_write_failure(const char *in_filename,
-    const char *out_filename, int err, char *err_info, uint32_t framenum,
+    const char *out_filename, int err, char *err_info, uint64_t framenum,
     int file_type_subtype);
 
 /*
